# Massively speed up zero_init workgroup memory tests on Android