@toji 2023-07-24
Not really familiar with the `SubcaseBatchState` but I was wondering if we had a mechanism for this already. I'll look into that.