@toji 2023-07-25
Updated this PR in light of discovering https://bugs.chromium.org/p/dawn/issues/detail?id=1933

Now it simply uses an explicit layout rather than `'auto'`, which should be a less disruptive change overall. I think there's still some benefit to be had from doing some explicit caching on the CTS side as way of helping other implementations and nudging people towards best practices, but it doesn't have to be handled here.